requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-crispy-forms==1.11.2
django-phonenumber-field==5.1.0
django-taggit==1.4.0
ediblepickle==1.1.3
folium==0.12.1
geographiclib==1.50
geoip2==4.1.0
geopy==2.1.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
maxminddb==2.0.3
multidict==5.1.0
numpy==1.20.2
phonenumbers==8.12.23
Pillow==8.2.0
pipenv==2018.11.26
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pytz==2021.1
requests==2.25.1
rope==0.19.0
signals==0.0.2
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==16.7.7
virtualenv-clone==0.5.3
yarl==1.6.3
